:construction: box/server/interface: Refactor structure.  Move Call related modules all down to call sub package.  Move Serdes to serdes sub package.
