+++ box/server/interface/call/control/_Access.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from dataclasses import field
 from time import time
 
-from ..._session import session_id
+from .._session import session_id
 from .._CallBase import CallBase
 from ._Hello import Processor
