+++ box/server/interface/call/control/_Command.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from dataclasses import field
 from typing import Any
 
-from ..._session import session_id
+from .._session import session_id
 from .._CallBase import CallBase
