+++ box/server/interface/call/control/_Ping.py
@@ -5,7 +5,7 @@ from dataclasses import field
 from math import nan
 from time import time
 
-from ..._session import session_id
+from .._session import session_id
 from .._CallBase import CallBase
