+++ box/server/interface/serdes/__init__.py
@@ -2,6 +2,4 @@
 
 from ._serdes import SerializeException, DeserializeException
 from ._serdes import serialize, deserialize
-
-from ._session import session_id, Sessions
-
+from ._serdes import ENCODING, IDX_BYTELEN, LEN_BYTELEN, BYTEORDER
